Clean-up of comments, etc.

+++ src/main/java/tungsten/types/vector/impl/RealVector.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright © 2018 Robert Poole <a href="mailto:Tarquin.AZ@gmail.com">Tarquin.AZ@gmail.com</a>.
+ * Copyright © 2018 Robert Poole <Tarquin.AZ@gmail.com>.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -97,7 +97,7 @@ public class RealVector implements Vector<RealType> {
 
     @Override
     public long length() {
-        return (long) elements.size();
+        return elements.size();
     }
 
     @Override
